+++ R/insertErrorTab.R
@@ -3,6 +3,7 @@
 #' @return text of the error list formated as an HTML page
 #' @param errorLst list of errors and changes made by \code{qcStudbook}
 #' @param pedigreeFileName name of file provided by user on Input tab
+#' @importFrom htmlTable htmlTable
 #' @importFrom stringi stri_c
 #' @importFrom stringi stri_split_regex
 #' @export
@@ -10,6 +11,7 @@ insertErrorTab <- function(errorLst, pedigreeFileName) {
   text <- summary(errorLst)
   if (checkChangedColsLst(errorLst$changedCols)) {
     colsChangedTxt <- "and Changes to Pedigree Column Names "
+
   } else {
     colsChangedTxt <- ""
   }
@@ -25,12 +27,7 @@ insertErrorTab <- function(errorLst, pedigreeFileName) {
                       line, "</li>\n")
   }
   if (nrow(text$sp) > 0) {
-    lines <- dataframe2string(text$sp, row.names = FALSE, digits = 2)
-    lines <- stri_split_lines(lines)[[1]]
-    newText <- stri_c(newText, "\n<h3>Questionable Parents Because of ",
-                      "Suspicious Dates</h3><pre>\n")
-    for (line in lines)
-      newText <- stri_c(newText, "<p>", line, "</p>")
+    newText <- stri_c(newText, htmlTable(text$sp))
   }
   newText <- stri_c(newText, "</ul>\n</p>\n</pre>")
   newText